Lombok deleted, DTO logic refactored, SQL data modified

+++ src/main/java/com/customerapp/management/repository/CustomerRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
 public interface CustomerRepository extends JpaRepository<Customer, Long>
 {
     @Query("SELECT AVG(c.age) FROM Customer c")
-    Double findAverageAge();
+    Float findAverageAge();
 }
